feat(generator): add types a few options and create a cli
