+++ .gitignore
@@ -1,2 +1,4 @@
 # IDE
 .idea
+dist
+node_modules
