+++ cli.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import * as gen from "./generator.js";
+
+console.log(gen.generateAsPeopleName("-", false));
