+++ tsconfig.json
@@ -1,0 +1,34 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "display": "Base TS Config",
+  "include": ["**/*", "**/*.json"],
+  "exclude": ["dist", ".next", "build", "node_modules", "**/__coverage__/**"],
+  "compilerOptions": {
+    "target": "es6",
+    "lib": ["esnext"],
+    "downlevelIteration": true,
+    "importHelpers": true,
+    "allowJs": true,
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "jsx": "react-jsx",
+    "noFallthroughCasesInSwitch": true,
+    "experimentalDecorators": true,
+    "incremental": true,
+    "composite": true,
+    "outDir": "dist"
+  },
+  "watchOptions": {
+    "watchFile": "useFsEvents",
+    "watchDirectory": "useFsEvents",
+    "fallbackPolling": "dynamicPriority",
+    "excludeDirectories": ["**/node_modules"]
+  }
+}
